# internal/httpserve/specs/openlane.openapi.json
{
  "components": {
...
    "schemas": {
...
      "ErrorReply": {
        "properties": {
          "error": {
            "description": "The error message if the request was unsuccessful",
            "type": "string"
          },
          "error_code": {
            "description": "The error code if the request was unsuccessful",
            "type": "string"
          },
          "success": {
            "description": "Whether or not the request was successful or not",
            "type": "boolean"
          },
          "unverified": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "ResetPasswordReply": {
        "properties": {
          "error": {
            "description": "The error message if the request was unsuccessful",
            "type": "string"
          },
          "error_code": {
            "description": "The error code if the request was unsuccessful",
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "description": "Whether or not the request was successful or not",
            "type": "boolean"
          },
          "unverified": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "ResetPasswordRequest": {
        "properties": {
          "password": {
            "description": "The new password to be used for authentication",
            "type": "string"
          },
          "token": {
            "description": "The token to be used to reset the password, token is sent via email",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "UnsubscribeReply": {
        "properties": {
          "error": {
            "description": "The error message if the request was unsuccessful",
            "type": "string"
          },
          "error_code": {
            "description": "The error code if the request was unsuccessful",
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "description": "Whether or not the request was successful or not",
            "type": "boolean"
          },
          "unverified": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "UnsubscribeRequest": {
        "properties": {
          "Token": {
            "description": "The token to be used to unsubscribe, token is sent via email",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Bearer token authentication using API tokens.\nThe token identifies the organization and grants permissions to manage\nusers and groups within that organization's scope.\n",
        "scheme": "bearer",
        "type": "http"
      },
      "apiKey": {
        "description": "API Key authentication, the key must be a valid API key",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "bearer": {
        "bearerFormat": "JWT",
        "description": "Bearer authentication, the token must be a valid API token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
...
  "info": {
    "contact": {
      "email": "support@theopenlane.io",
      "name": "Openlane",
      "url": "https://www.theopenlane.io"
    },
    "description": "Openlane's API services are designed to provide a simple and easy to use interface for interacting with the Openlane platform. This API is designed to be used by both internal and external clients to interact with the Openlane platform.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "Openlane OpenAPI 3.1.1 Specifications",
    "version": "v0.0.0-dev"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/v1/password-reset": {
      "post": {
        "description": "ResetPassword allows the user (after requesting a password reset) to set a new password - the password reset token needs to be set in the request and not expired. If the request is successful, a confirmation of the reset is sent to the user and a 200 StatusOK is returned",
        "operationId": "ResetPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "ResetPasswordSuccessRequest": {
                  "value": {
                    "password": "mitb!",
                    "token": "token"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "description": "Request body"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "ResetPasswordSuccessResponse": {
                    "value": {
                      "message": "Password has been reset",
                      "success": true
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordReply"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "bad request",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Bad Request"
          },
          "429": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "too many requests",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "internal server error",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [],
        "summary": "ResetPassword",
        "tags": [
          "Account Management"
        ]
      }
    },
...
    "/v1/unsubscribe": {
      "post": {
        "description": "Unsubscribe from communications",
        "operationId": "Unsubscribe",
        "parameters": [
          {
            "description": "The token to be used to unsubscribe, token is sent via email",
            "in": "query",
            "name": "token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "UnsubscribeRequest": {
                  "value": {
                    "Token": "token"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/UnsubscribeRequest"
              }
            }
          },
          "description": "Request body"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "UnsubscribeResponse": {
                    "value": {
                      "message": "You have been unsubscribed and will no longer receive updates.",
                      "success": true
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/UnsubscribeReply"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "bad request",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Bad Request"
          },
          "429": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "too many requests",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "internal server error",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [],
        "summary": "Unsubscribe",
        "tags": [
          "Subscribers"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Openlane API Server",
      "url": "https://api.theopenlane.io"
    }
  ],
...
}